src-command/ingest/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
pymongo==4.16.0
kafka-python==2.3.0
pydantic==2.12.5
prometheus-client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
